+++ tests/Fixtures/Integration/Resources/Browser/BrowserMergeEditPost.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Lenorix\FilamentAutosave\Tests\Fixtures\Integration\Resources\Browser;
+
+use Filament\Resources\Pages\EditRecord;
+use Lenorix\FilamentAutosave\HasAutosave;
+
+class BrowserMergeEditPost extends EditRecord
+{
+    use HasAutosave;
+
+    protected static string $resource = BrowserMergePostResource::class;
+
+    /** @return array<int, string> */
+    protected function autosaveMergeFields(): ?array
+    {
+        return ['body'];
+    }
+}